Add model checking file for DistLog and move to more abstract representation of 'chosen' set
Decided to move to a more abstract representation of 'chosen' as a set
of <<index, value>> pairs instead of an explicit sequence, so that we
could actually represent the case where two different values were chosen
at the same index. Right now I think the refinement mapping between
'DistLog' and 'Log' is broken because 'DistLog' can currently mark
entries as "chosen" out of sequential order, which isn't allowed by
'Log'. I don't think this is a fundamental problem, but I will have to
either alter the behavior of 'DistLog' or alter the refinement mapping
to address this mismatch.