Initial implementation of ApplyEvents for a couple of event types
The 'BecomeLeader' implementation is still incomplete and I need to
write all the other action types. I might also need to change what data
is stored with each event to keep the right amount of data to execute
the event during "replay". Right now I store the entire pre and post
image of the state before/after the execution of the event.